worked on admin dashboard

+++ routes/web.php
@@ -16,9 +16,8 @@ use Illuminate\Support\Facades\Route;
 
 
 
-Route::get('/dashboard', function () {
-    return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+Route::get('/dashboard',[PagesController::class, 'dashboard'])->middleware(['auth', 'verified'])->name('dashboard');
+
 
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
@@ -31,13 +30,15 @@ require __DIR__.'/auth.php';
 //pages
 Route::get('/', [PagesController::class,'index']);
 Route::get('productpage', [PagesController::class,'productpage']);
-Route::get('cartpage', [PagesController::class,'cartpage']);
+Route::get('cartpage', [PagesController::class,'cartpage'])->middleware(['auth', 'verified']);
 Route::get('contact', [PagesController::class,'contact']);
-Route::get('checkoutpage', [PagesController::class,'checkoutpage']);
+Route::get('checkoutpage', [PagesController::class,'checkoutpage'])->middleware(['auth', 'verified']);
 Route::get('singleproduct/{id}', [PagesController::class,'singleproduct']);
+Route::get('/dashboard',[PagesController::class,'redirect']);
+
 
 // user
-Route::get('getAllUser',[UserController::class, 'getAllUser']);
+Route::get('getAllUser',[UserController::class, 'getAllUser'])->middleware(['auth', 'verified']);
 Route::get('getAUser/id',[UserController::class, 'getAUser']);
 Route::patch('updateUser/id', [UserController::class, 'updateUser']);
 Route::delete('deleteAUser/id', [UserController::class, 'deleteAUser']);
@@ -69,24 +70,23 @@ Route::patch('updateProduct/id', [ProductController::class, 'updateProduct']);
 Route::delete('deleteAProduct/id', [ProductController::class, 'deleteAProduct']);
 Route::post('createProduct', [ProductController::class, 'createProduct']);
 //cart
-Route::post('/addcart/{id}', [CartController::class, 'addcart']);
-Route::get('removecart/{id}', [CartController::class, 'removecart']);
-Route::delete('deletecart/{id}', [CartController::class, 'deletecart']);
-Route::get('addquantity/{id}',[CartController::class, 'addquantity']);
-Route::get('removequantity/{id}',[CartController::class, 'removequantity']);
+Route::post('/addcart/{id}', [CartController::class, 'addcart'])->middleware(['auth', 'verified']);
+Route::get('removecart/{id}', [CartController::class, 'removecart'])->middleware(['auth', 'verified']);
+Route::delete('deletecart/{id}', [CartController::class, 'deletecart'])->middleware(['auth', 'verified']);
+Route::get('addquantity/{id}',[CartController::class, 'addquantity'])->middleware(['auth', 'verified']);
+Route::get('removequantity/{id}',[CartController::class, 'removequantity'])->middleware(['auth', 'verified']);
 
 //Payment
-Route::post('payondelivery',[CartController::class,'payondelivery']);
+Route::post('payondelivery',[CartController::class,'payondelivery'])->middleware(['auth', 'verified']);
 Route::post('pay',[PaymentController::class,'redirectToGateway'])->name('pay');
 Route::get('paymentcallback', [PaymentController::class, 'handleGatewayCallback'])->name('name.callback');
 //admin panel
-Route::get('indexadmin', [AdminController::class, 'indexadmin']);
-Route::get('product',[AdminController::class,'product']);
-Route::get('addproduct', [AdminController::class,'addproduct']);
-Route::get('order', [AdminController::class, 'order']);
+Route::get('indexadmin', [AdminController::class, 'indexadmin'])->middleware(['auth', 'verified']);
+Route::get('product',[AdminController::class,'product'])->middleware(['auth', 'verified']);
+Route::get('addproduct', [AdminController::class,'addproduct'])->middleware(['auth', 'verified']);
 Route::get('message', [AdminController::class, 'message']);
 Route::get('team', [AdminController::class, 'team']);
-Route::get('orders', [AdminController::class, 'order']);
+Route::get('orders', [AdminController::class, 'order'])->middleware(['auth', 'verified']);
 Route::post('addmessage', [AdminController::class, 'addmessage']);
 Route::get('category', [AdminController::class, 'category']);
 Route::get('brand', [AdminController::class, 'brand']);
